+++ Tests/Utility/ApiLogUtilityTest.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace Chaplean\Bundle\ApiClientBundle\Utility;
+namespace Tests\Chaplean\Bundle\ApiClientBundle\Utility;
 
 use Chaplean\Bundle\ApiClientBundle\Api\Response\Success\PlainResponse;
 use Chaplean\Bundle\ApiClientBundle\Entity\ApiLog;
 use Chaplean\Bundle\ApiClientBundle\Entity\ApiMethodType;
 use Chaplean\Bundle\ApiClientBundle\Entity\ApiStatusCodeType;
 use Chaplean\Bundle\ApiClientBundle\Query\ApiLogQuery;
+use Chaplean\Bundle\ApiClientBundle\Utility\ApiLogUtility;
 use Doctrine\Bundle\DoctrineBundle\Registry;
 use Doctrine\ORM\AbstractQuery;
-use Doctrine\ORM\EntityManager;
+use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\EntityRepository;
-use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
 use Mockery\Adapter\Phpunit\MockeryTestCase;
 
@@ -25,12 +25,34 @@ use Mockery\Adapter\Phpunit\MockeryTestCase;
 class ApiLogUtilityTest extends MockeryTestCase
 {
     /**
+     * @var ApiLogQuery|\Mockery\MockInterface
+     */
+    private $apiLogQuery;
+
+    /**
+     * @var Registry|\Mockery\MockInterface
+     */
+    private $registry;
+
+    /**
+     * @var EntityManagerInterface|\Mockery\MockInterface
+     */
+    private $em;
+
+    protected function setUp()
+    {
+        $this->apiLogQuery = \Mockery::mock(ApiLogQuery::class);
+        $this->em = \Mockery::mock(EntityManagerInterface::class);
+        $this->registry = \Mockery::mock(Registry::class);
+    }
+
+    /**
      * @covers \Chaplean\Bundle\ApiClientBundle\Utility\ApiLogUtility::__construct()
      * @covers \Chaplean\Bundle\ApiClientBundle\Utility\ApiLogUtility::logResponse()
      *
      * @return void
      */
-    public function testLogRequestInDatabaseIfEnabled()
+    public function testLogRequestInDatabase()
     {
         $methodType = new ApiMethodType();
         $apiMethodRepo = \Mockery::mock(EntityRepository::class);
@@ -40,7 +62,7 @@ class ApiLogUtilityTest extends MockeryTestCase
         $apiStatusCodeRepo = \Mockery::mock(EntityRepository::class);
         $apiStatusCodeRepo->shouldReceive('findOneBy')->once()->with(['code' => 200])->andReturn($statusCodeType);
 
-        $em = \Mockery::mock(EntityManager::class);
+        $em = \Mockery::mock(EntityManagerInterface::class);
         $em->shouldReceive('getRepository')->once()->with(ApiMethodType::class)->andReturn($apiMethodRepo);
 
         $em->shouldReceive('getRepository')->once()->with(ApiStatusCodeType::class)->andReturn($apiStatusCodeRepo);
@@ -53,7 +75,7 @@ class ApiLogUtilityTest extends MockeryTestCase
         $registry->shouldReceive('getManager')->once()->andReturn($em);
 
         $config = [
-            'enable_database_logging' => true,
+            'enable_database_logging' => null // true value will be convert to null on configuration normalization
         ];
 
         $utility = new ApiLogUtility($config, $apiLogQuery, $registry);
@@ -76,7 +98,7 @@ class ApiLogUtilityTest extends MockeryTestCase
         $apiStatusCodeRepo = \Mockery::mock(EntityRepository::class);
         $apiStatusCodeRepo->shouldReceive('findOneBy')->once()->with(['code' => 418])->andReturn($statusCodeType);
 
-        $em = \Mockery::mock(EntityManager::class);
+        $em = \Mockery::mock(EntityManagerInterface::class);
         $em->shouldReceive('getRepository')->once()->with(ApiMethodType::class)->andReturn($apiMethodRepo);
 
         $em->shouldReceive('getRepository')->once()->with(ApiStatusCodeType::class)->andReturn($apiStatusCodeRepo);
@@ -89,35 +111,12 @@ class ApiLogUtilityTest extends MockeryTestCase
         $registry->shouldReceive('getManager')->once()->andReturn($em);
 
         $config = [
-            'enable_database_logging' => true,
+            'enable_database_logging' => null, // true value will be convert to null on configuration normalization
         ];
 
         $utility = new ApiLogUtility($config, $apiLogQuery, $registry);
 
         $utility->logResponse(new PlainResponse(new Response(418, [], ''), 'get', 'url', []));
-    }
-
-    /**
-     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\ApiLogUtility::__construct()
-     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\ApiLogUtility::logResponse()
-     *
-     * @return void
-     */
-    public function testDontLogRequestInDatabaseIfDisabled()
-    {
-        $em = \Mockery::mock(EntityManager::class);
-
-        $apiLogQuery = \Mockery::mock(ApiLogQuery::class);
-        $registry = \Mockery::mock(Registry::class);
-        $registry->shouldReceive('getManager')->once()->andReturn($em);
-
-        $config = [
-            'enable_database_logging' => false,
-        ];
-
-        $utility = new ApiLogUtility($config, $apiLogQuery, $registry);
-
-        $utility->logResponse(new PlainResponse(new Response(200, [], ''), 'get', 'url', []));
     }
 
     /**
@@ -133,7 +132,7 @@ class ApiLogUtilityTest extends MockeryTestCase
         $apiLogQuery = \Mockery::mock(ApiLogQuery::class);
 
         $config = [
-            'enable_database_logging' => true,
+            'enable_database_logging' => null,
         ];
 
         new ApiLogUtility($config, $apiLogQuery);
@@ -152,7 +151,6 @@ class ApiLogUtilityTest extends MockeryTestCase
         $apiLogQuery = \Mockery::mock(ApiLogQuery::class);
 
         $config = [
-            'enable_database_logging' => false,
         ];
 
         new ApiLogUtility($config, $apiLogQuery);
@@ -167,7 +165,7 @@ class ApiLogUtilityTest extends MockeryTestCase
     public function testDeleteMostRecentThan()
     {
         $date = new \DateTime();
-        $em = \Mockery::mock(EntityManager::class);
+        $em = \Mockery::mock(EntityManagerInterface::class);
         $querySearch = \Mockery::mock(AbstractQuery::class);
         $queryDelete = \Mockery::mock(AbstractQuery::class);
         $apiLogQuery = \Mockery::mock(ApiLogQuery::class);
@@ -201,12 +199,12 @@ class ApiLogUtilityTest extends MockeryTestCase
     {
         $apiLogQuery = \Mockery::mock(ApiLogQuery::class);
         $registry = \Mockery::mock(Registry::class);
-        $em = \Mockery::mock(EntityManager::class);
+        $em = \Mockery::mock(EntityManagerInterface::class);
         $repository = \Mockery::mock(EntityRepository::class);
         $log = new ApiLog();
 
         $config = [
-            'enable_database_logging' => true,
+            'enable_database_logging' =>  null, // true value will be convert to null on configuration normalization,
         ];
 
         $registry->shouldReceive('getManager')->once()->andReturn($em);
@@ -233,7 +231,7 @@ class ApiLogUtilityTest extends MockeryTestCase
         $apiStatusCodeRepo = \Mockery::mock(EntityRepository::class);
         $apiStatusCodeRepo->shouldReceive('findOneBy')->once()->with(['code' => 200])->andReturn($statusCodeType);
 
-        $em = \Mockery::mock(EntityManager::class);
+        $em = \Mockery::mock(EntityManagerInterface::class);
         $em->shouldReceive('getRepository')->once()->with(ApiMethodType::class)->andReturn($apiMethodRepo);
 
         $em->shouldReceive('getRepository')->once()->with(ApiStatusCodeType::class)->andReturn($apiStatusCodeRepo);
@@ -246,7 +244,7 @@ class ApiLogUtilityTest extends MockeryTestCase
         $registry->shouldReceive('getManager')->once()->andReturn($em);
 
         $config = [
-            'enable_database_logging' => true,
+            'enable_database_logging' =>  null, // true value will be convert to null on configuration normalization,
         ];
 
         $response = new PlainResponse(new Response(200, [], ''), 'get', 'url', []);
@@ -257,5 +255,167 @@ class ApiLogUtilityTest extends MockeryTestCase
         $em->shouldNotReceive('getRepository')->with(ApiLog::class);
 
         $this->assertNotNull($utility->getLogByUuid($response->getUuid()));
+    }
+
+    /**
+     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\ApiLogUtility::__construct
+     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\ApiLogUtility::isEnabledLoggingFor
+     *
+     * @return void
+     */
+    public function testLoggingIsEnabledExplicitDefinition()
+    {
+        $this->registry->shouldReceive('getManager')->once()->andReturn($this->em);
+
+        $utility = new ApiLogUtility(
+            [
+                'enable_database_logging' => [
+                    'type'     => 'inclusive',
+                    'elements' => [
+                        'bar_api'
+                    ]
+                ]
+            ], 
+            $this->apiLogQuery,
+            $this->registry
+        );
+
+        $this->assertTrue($utility->isEnabledLoggingFor('bar_api'));
+    }
+
+    /**
+     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\ApiLogUtility::isEnabledLoggingFor
+     *
+     * @return void
+     */
+    public function testLoggingIsEnabledTildDefinition()
+    {
+        $this->registry->shouldReceive('getManager')->once()->andReturn($this->em);
+
+        $utility = new ApiLogUtility(
+            [
+                'enable_database_logging' => null
+            ],
+            $this->apiLogQuery,
+            $this->registry
+        );
+
+        $this->assertTrue($utility->isEnabledLoggingFor('bar_api'));
+    }
+
+    /**
+     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\ApiLogUtility::isEnabledLoggingFor
+     *
+     * @return void
+     */
+    public function testLoggingIsEnabledExclusiveDefinition()
+    {
+        $this->registry->shouldReceive('getManager')->once()->andReturn($this->em);
+
+        $utility = new ApiLogUtility(
+            [
+                'enable_database_logging' => [
+                    'type' => 'exclusive',
+                    'elements' => [
+                        'foo_api'
+                    ]
+                ]
+            ],
+            $this->apiLogQuery,
+            $this->registry
+        );
+
+        $this->assertTrue($utility->isEnabledLoggingFor('bar_api'));
+    }
+
+    /**
+     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\ApiLogUtility::isEnabledLoggingFor
+     *
+     * @return void
+     */
+    public function testLoggingIsDisabledExplicitDefinition()
+    {
+        $this->registry->shouldReceive('getManager')->once()->andReturn($this->em);
+
+        $utility = new ApiLogUtility(
+            [
+                'enable_database_logging' => [
+                    'type' => 'exclusive',
+                    'elements' => [
+                        'bar_api'
+                    ]
+                ]
+            ],
+            $this->apiLogQuery,
+            $this->registry
+        );
+
+        $this->assertFalse($utility->isEnabledLoggingFor('bar_api'));
+    }
+
+    /**
+     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\ApiLogUtility::isEnabledLoggingFor
+     *
+     * @return void
+     */
+    public function testLoggingIsDisabledByDefault()
+    {
+        $this->registry->shouldReceive('getManager')->once()->andReturn($this->em);
+
+        $utility = new ApiLogUtility(
+            [],
+            $this->apiLogQuery,
+            $this->registry
+        );
+
+        $this->assertFalse($utility->isEnabledLoggingFor('bar_api'));
+    }
+
+    /**
+     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\ApiLogUtility::isEnabledLoggingFor
+     *
+     * @return void
+     */
+    public function testLoggingIsDisabledNotDefineApiName()
+    {
+        $this->registry->shouldReceive('getManager')->once()->andReturn($this->em);
+
+        $utility = new ApiLogUtility(
+            [
+                'enable_database_logging' => [
+                    'type' => 'inclusive',
+                    'elements' => [
+                        'foo_api'
+                    ]
+                ]
+            ],
+            $this->apiLogQuery,
+            $this->registry
+        );
+
+        $this->assertFalse($utility->isEnabledLoggingFor('bar_api'));
+    }
+
+    /**
+     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\ApiLogUtility::logResponse
+     *
+     * @return void
+     */
+    public function testLogResponseWithDisabledApi()
+    {
+        $this->registry->shouldReceive('getManager')->once()->andReturn($this->em);
+        $this->em->shouldNotReceive('persist');
+
+        $config = [
+            'enable_database_logging' => [
+                'type'     => 'exclusive',
+                'elements' => [
+                    'foo_api'
+                ]
+            ]
+        ];
+
+        $utility = new ApiLogUtility($config, $this->apiLogQuery, $this->registry);
+        $utility->logResponse(new PlainResponse(new Response(200, [], ''), 'get', 'url', []), 'foo_api');
     }
 }