+++ Tests/Utility/EmailUtilityTest.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace Chaplean\Bundle\ApiClientBundle\Utility;
+namespace Tests\Chaplean\Bundle\ApiClientBundle\Utility;
 
 use Chaplean\Bundle\ApiClientBundle\Api\Response\Success\PlainResponse;
 use Chaplean\Bundle\ApiClientBundle\Tests\Resources\DataProviderTrait;
+use Chaplean\Bundle\ApiClientBundle\Utility\EmailUtility;
 use GuzzleHttp\Psr7\Response;
 use Mockery\Adapter\Phpunit\MockeryTestCase;
 use Symfony\Bridge\Twig\TwigEngine;
@@ -19,6 +20,21 @@ use Symfony\Component\Translation\TranslatorInterface;
 class EmailUtilityTest extends MockeryTestCase
 {
     use DataProviderTrait;
+
+    /**
+     * @var \Mockery\MockInterface
+     */
+    private $mailer;
+
+    /**
+     * @var \Mockery\MockInterface
+     */
+    private $translator;
+
+    /**
+     * @var \Mockery\MockInterface
+     */
+    private $templating;
 
     /**
      * @return void
@@ -51,7 +67,7 @@ class EmailUtilityTest extends MockeryTestCase
         $this->templating = \Mockery::mock(TwigEngine::class);
 
         $config = [
-            'enable_email_logging' => true,
+            'enable_email_logging' => null,
             'email_logging'        => [
                 'codes_listened' => $config,
                 'address_from'   => 'test@example.com',
@@ -78,7 +94,7 @@ class EmailUtilityTest extends MockeryTestCase
         $this->mailer->shouldReceive('send')->once();
 
         $config = [
-            'enable_email_logging' => true,
+            'enable_email_logging' => null,
             'email_logging'        => [
                 'codes_listened' => ['0', '4XX', '5XX'],
                 'address_from'   => 'test@example.com',
@@ -101,7 +117,7 @@ class EmailUtilityTest extends MockeryTestCase
         $this->mailer->shouldNotReceive('send');
 
         $config = [
-            'enable_email_logging' => true,
+            'enable_email_logging' => null,
             'email_logging'        => [
                 'codes_listened' => ['0', '4XX', '5XX'],
                 'address_from'   => 'test@example.com',
@@ -115,29 +131,6 @@ class EmailUtilityTest extends MockeryTestCase
 
     /**
      * @covers \Chaplean\Bundle\ApiClientBundle\Utility\EmailUtility::__construct()
-     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\EmailUtility::sendRequestExecutedNotificationEmail()
-     *
-     * @return void
-     */
-    public function testDontSendMailIfDisabled()
-    {
-        $this->mailer->shouldNotReceive('send');
-
-        $config = [
-            'enable_email_logging' => false,
-            'email_logging'        => [
-                'codes_listened' => ['0', '4XX', '5XX'],
-                'address_from'   => 'test@example.com',
-                'address_to'     => 'test@example.com'
-            ],
-        ];
-
-        $utility = new EmailUtility($config, $this->mailer, $this->translator, $this->templating);
-        $utility->sendRequestExecutedNotificationEmail(new PlainResponse(new Response(501, [], ''), 'get', 'url', []));
-    }
-
-    /**
-     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\EmailUtility::__construct()
      *
      * @expectedException \InvalidArgumentException
      * @expectedExceptionMessage Email logging is enabled, you must register the mailer, translator and twig services
@@ -147,7 +140,7 @@ class EmailUtilityTest extends MockeryTestCase
     public function testConstructFailsIfConfigEnablesLoggingWithoutTheRequiredServices()
     {
         $config = [
-            'enable_email_logging' => true,
+            'enable_email_logging' => null,
             'email_logging'        => [
                 'codes_listened' => ['0', '4XX', '5XX'],
                 'address_from'   => 'test@example.com',
@@ -159,25 +152,157 @@ class EmailUtilityTest extends MockeryTestCase
     }
 
     /**
-     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\EmailUtility::__construct()
-     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\EmailUtility::sendRequestExecutedNotificationEmail()
+     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\EmailUtility::__construct
+     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\EmailUtility::isEnabledLoggingFor
      *
      * @return void
      */
-    public function testMissingServicesAreIgnoredIfLoggingIsDisabled()
+    public function testLoggingIsEnabledExplicitDefinition()
+    {
+        $utility = new EmailUtility(
+            [
+                'enable_email_logging' => [
+                    'type'     => 'inclusive',
+                    'elements' => [
+                        'bar_api'
+                    ]
+                ]
+            ],
+            $this->mailer,
+            $this->translator,
+            $this->templating
+        );
+
+        $this->assertTrue($utility->isEnabledLoggingFor('bar_api'));
+    }
+
+    /**
+     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\EmailUtility::isEnabledLoggingFor
+     *
+     * @return void
+     */
+    public function testLoggingIsEnabledTildDefinition()
+    {
+        $utility = new EmailUtility(
+            [
+                'enable_email_logging' => null
+            ],
+            $this->mailer,
+            $this->translator,
+            $this->templating
+        );
+
+        $this->assertTrue($utility->isEnabledLoggingFor('bar_api'));
+    }
+
+    /**
+     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\EmailUtility::isEnabledLoggingFor
+     *
+     * @return void
+     */
+    public function testLoggingIsEnabledExclusiveDefinition()
+    {
+        $utility = new EmailUtility(
+            [
+                'enable_email_logging' => [
+                    'type' => 'exclusive',
+                    'elements' => [
+                        'foo_api'
+                    ]
+                ]
+            ],
+            $this->mailer,
+            $this->translator,
+            $this->templating
+        );
+
+        $this->assertTrue($utility->isEnabledLoggingFor('bar_api'));
+    }
+
+    /**
+     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\EmailUtility::isEnabledLoggingFor
+     *
+     * @return void
+     */
+    public function testLoggingIsDisabledExplicitDefinition()
+    {
+        $utility = new EmailUtility(
+            [
+                'enable_email_logging' => [
+                    'type' => 'exclusive',
+                    'elements' => [
+                        'bar_api'
+                    ]
+                ]
+            ],
+            $this->mailer,
+            $this->translator,
+            $this->templating
+        );
+
+        $this->assertFalse($utility->isEnabledLoggingFor('bar_api'));
+    }
+
+    /**
+     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\EmailUtility::isEnabledLoggingFor
+     *
+     * @return void
+     */
+    public function testLoggingIsDisabledByDefault()
+    {
+        $utility = new EmailUtility(
+            [],
+            $this->mailer,
+            $this->translator,
+            $this->templating
+        );
+
+        $this->assertFalse($utility->isEnabledLoggingFor('bar_api'));
+    }
+
+    /**
+     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\EmailUtility::isEnabledLoggingFor
+     *
+     * @return void
+     */
+    public function testLoggingIsDisabledNotDefineApiName()
+    {
+        $utility = new EmailUtility(
+            [
+                'enable_email_logging' => [
+                    'type' => 'inclusive',
+                    'elements' => [
+                        'foo_api'
+                    ]
+                ]
+            ],
+            $this->mailer,
+            $this->translator,
+            $this->templating
+        );
+
+        $this->assertFalse($utility->isEnabledLoggingFor('bar_api'));
+    }
+
+    /**
+     * @covers \Chaplean\Bundle\ApiClientBundle\Utility\EmailUtility::sendRequestExecutedNotificationEmail
+     *
+     * @return void
+     */
+    public function testSendRequestExecutedNotificationEmailWithDisabledApi()
     {
         $this->mailer->shouldNotReceive('send');
 
         $config = [
-            'enable_email_logging' => false,
-            'email_logging'        => [
-                'codes_listened' => ['0', '4XX', '5XX'],
-                'address_from'   => 'test@example.com',
-                'address_to'     => 'test@example.com'
-            ],
+            'enable_email_logging' => [
+                'type'     => 'exclusive',
+                'elements' => [
+                    'foo_api'
+                ]
+            ]
         ];
 
-        $utility = new EmailUtility($config);
-        $utility->sendRequestExecutedNotificationEmail(new PlainResponse(new Response(501, [], ''), 'get', 'url', []));
+        $utility = new EmailUtility($config, $this->mailer, $this->translator, $this->templating);
+        $utility->sendRequestExecutedNotificationEmail(new PlainResponse(new Response(200, [], ''), 'get', 'url', []), 'foo_api');
     }
 }